Image y values should be ushort, not uchar. We may move to 640x480.

+++ ImagePassHandler/FieldEdgePass/fieldedgepass.cc
@@ -32,7 +32,7 @@ void FieldEdgePass::onPixel(uchar labelId, ushort x, ushort y)
   }
 }
 
-uchar FieldEdgePass::getEdgeYValue(ushort x) const
+ushort FieldEdgePass::getEdgeYValue(ushort x) const
 {
   assert(x < d_pixelWidth);
   assert(d_maxYByX[x] < d_pixelHeight);
